feat: allow informative schema checking (reported as new RSC-024 and RSC-025)
An `XMLValidator` instance can now be instantiated with a boolean
flag to tell if the unerlying schema is normative or not.

If a schema is not normative, the `XMLParser` uses an informative-only `ErrorHandler`,
that reports warnings as "RSC-024" (USAGE), and errors as "RSC-025" (USAGE).
Fatal errors are still reported as "RSC-016" (FATAL).