feat: no longer check `.xhtml` extension in EPUB 2.0.1
EPUB 2.0.1 does not specifically require XHTML content documents to have
the `.xhtml` extension.

EPUBCheck reported this as a warning (`HTM-014`), now suppressed.
The sibling `HTM-014a`, which checked the same for EPUB 3,  was already
suppressed.

Note: this change will only be effective when the checks of the `ctc`
package  are disabled.