feat: disable some cases of redundant missing resources errors
The systematic reporting of `RSC-001` in the `ctc` package caused the error
to be raised even when the issue was already reported specifically earlier
(which is the case for key files like `container.xml` and the package document).

The removal of the `ctc` package will effectively remove this redundant
reporting.

Note: this change will only be effective when the checks of the `ctc`
package  are disabled.