feat: no longer check that navigation links to all spine items
EPUB does not require that the navigation document (or NCX in EPUB 2.0.1) links
to all the documents listed in the spine.

EPUBCheck reported this as USAGE messages (`OPF-059` and `OPF-058`),
which are now suppressed.

Note: this change will only be effective when the checks of the `ctc`
package  are disabled.