fix: allow HTML `script` element in EPUB 2.0.1
EPUB 2.0.1 does not strictly forbids scripts in HTML. It does say that a reading
system should not execute the script, but there is no conformance statement
on the content side.

This was reported as `SCP-004`, now suppressed.

Note: this change will only be effective when the checks of the `ctc`
package  are disabled.