refactor: major rework of URL handling
EPUBCheck now uses URL records internally for everything related
to paths and URLs handling.

We use the Galimatias library which is very good, close to the URL
standard, and also used in Validator.nu.

This change also include many other improvements to the code and
internal APIs used in EPUBCheck checkers.